[procon test] Run test src/atcoder/abc127/c.cc

+++ src/atcoder/abc127/c.cc
@@ -821,5 +821,5 @@ void body() {
         max = std::min(max, LR.second);
     }
 
-    cout << std::max(0, max - min + 1) << endl;
+    cout << std::max(i64(0), max - min + 1) << endl;
 }
